Remove the AMD bundle loader and the esmModuleLoader flag (src)
The ESM module-record loader has been the production default since the
flag flip; this removes the legacy AMD path and the flag entirely:

- runner: drop esm-loader-config.ts, the experimental.esmModuleLoader
  option, Engine.compile()/evaluate() (AMD bundle compile/eval), the
  CompiledBundleValidator + bundle-preflight, and the AMD entry points
  of compiled-bundle-verifier (classifyModuleItems core stays — the ESM
  body verifier builds on it; the canonical tslib-helper recognition
  moves to sandbox/tslib-helpers.ts). PatternManager now compiles
  unconditionally through the record-graph path.
- AMD-only identity plumbing: the CT-1665 verified-binding-candidate
  registrar (builder hook + engine capture) retires with the AMD eval
  path, as planned — the ESM loader registers non-exported trusted
  bindings through the transformer's __cfReg registration sink.
- compilation cache (AMD bundle cache): remove CachedCompiler and its
  memory/fs/idb storage backends + git-fingerprint, the runtime
  cachedCompiler option, toolshed/runtime-client construction, and the
  COMPILATION_CACHE_* env plumbing. The ESM path's content-addressed
  cell cache (cell-cache.ts) is unaffected. The worker-URL ?v=buildHash
  cache-buster in lib-shell is kept, now unconditional.
- js-compiler: remove the AMD bundler (typescript/bundler/) and the
  bundled compile()/resolveAndCompile(); compileToModules (per-module
  CommonJS) is the only emit path and now honors getTransformedProgram.
  Base compiler options emit CommonJS directly.
- cli: cf check/dev, the pattern test runner, and the multi-user test
  worker now compile through the record-graph path; --output writes the
  concatenated per-module compiled bodies; --filename is gone (it only
  named the AMD bundle). compileToRecordGraph now forwards
  getTransformedProgram and wires OpaqueRefErrorTransformer, restoring
  cf check's --show-transformed and simplified type-error hints.
- shell / lib-shell / runtime-client / toolshed: drop the flag and
  compilation-cache-client plumbing (defines, env parsing, protocol
  fields).
