--- packages/js-compiler/typescript/bundler/amd-loader.ts
@@ -1,185 +1,0 @@
-// deno-lint-ignore-file no-explicit-any
-
-// This function is a minimal AMD implementation that gets
-// stringified via `.toString()` (and "minified", stripping new lines)
-// in order to load the bundle generated by the TypeScript compiler.
-export function getAMDLoader(
-  internalHooks: {
-    define?: (
-      id: string,
-      dependencies: string[],
-      factorySource: string,
-    ) => void;
-    require?: (dependencies: string[] | string) => void;
-  } = {},
-) {
-  interface ModuleDefinition {
-    dependencies: string[];
-    factory: (...deps: any[]) => any;
-    exports?: any;
-    resolved?: boolean;
-  }
-
-  class AMDLoader {
-    private modules = new Map<string, ModuleDefinition>();
-    private loading = new Set<string>();
-
-    define(
-      id: string,
-      dependencies: string[],
-      factory: (...deps: any[]) => any,
-    ): void;
-    define(dependencies: string[], factory: (...deps: any[]) => any): void;
-    define(factory: () => any): void;
-    define(
-      idOrDepsOrFactory: string | string[] | (() => any),
-      depsOrFactory?: string[] | ((...deps: any[]) => any),
-      factory?: (...deps: any[]) => any,
-    ): void {
-      let id: string;
-      let dependencies: string[];
-      let moduleFactory: (...deps: any[]) => any;
-
-      if (typeof idOrDepsOrFactory === "string") {
-        id = idOrDepsOrFactory;
-        dependencies = depsOrFactory as string[];
-        moduleFactory = factory!;
-      } else if (Array.isArray(idOrDepsOrFactory)) {
-        id = this.generateAnonymousId();
-        dependencies = idOrDepsOrFactory;
-        moduleFactory = depsOrFactory as (...deps: any[]) => any;
-      } else {
-        id = this.generateAnonymousId();
-        dependencies = [];
-        moduleFactory = idOrDepsOrFactory;
-      }
-
-      if (this.modules.has(id)) {
-        throw new Error(`Module '${id}' is already defined`);
-      }
-
-      this.modules.set(id, {
-        dependencies,
-        factory: moduleFactory,
-        resolved: false,
-      });
-      internalHooks.define?.(
-        id,
-        dependencies,
-        Function.prototype.toString.call(moduleFactory),
-      );
-    }
-
-    require(
-      dependencies: string[],
-      callback: (...modules: any[]) => void,
-    ): void;
-    require(id: string): any;
-    require(
-      depsOrId: string[] | string,
-      callback?: (...modules: any[]) => void,
-    ): any {
-      if (typeof depsOrId === "string") {
-        internalHooks.require?.(depsOrId);
-        return this.resolveModule(depsOrId);
-      } else {
-        internalHooks.require?.(depsOrId);
-        const resolvedModules = depsOrId.map((dep) => this.resolveModule(dep));
-        callback?.(...resolvedModules);
-      }
-    }
-
-    private authoredRequire(): never {
-      throw new Error("Authored AMD require() is unavailable in SES mode");
-    }
-
-    private freezeResolvedExports<T>(value: T): T {
-      if (
-        value !== null &&
-        (typeof value === "object" || typeof value === "function")
-      ) {
-        Object.freeze(value);
-      }
-      return value;
-    }
-
-    private resolveModule(id: string): any {
-      const module = this.modules.get(id);
-      if (!module) {
-        throw new Error(`Module '${id}' not found`);
-      }
-
-      if (module.resolved) {
-        return module.exports;
-      }
-
-      if (this.loading.has(id)) {
-        throw new Error(`Circular dependency detected: ${id}`);
-      }
-
-      this.loading.add(id);
-
-      try {
-        const resolvedDeps = module.dependencies.map((dep) => {
-          if (dep === "require") return this.authoredRequire.bind(this);
-          if (dep === "exports") return {};
-          if (dep === "module") return { exports: {} };
-          return this.resolveModule(dep);
-        });
-
-        const result = module.factory(...resolvedDeps);
-
-        const exportsIndex = module.dependencies.indexOf("exports");
-        const moduleIndex = module.dependencies.indexOf("module");
-        const exportsObject = exportsIndex >= 0 &&
-            typeof resolvedDeps[exportsIndex] === "object"
-          ? resolvedDeps[exportsIndex]
-          : undefined;
-        const moduleObject = moduleIndex >= 0 &&
-            resolvedDeps[moduleIndex] &&
-            typeof resolvedDeps[moduleIndex] === "object"
-          ? resolvedDeps[moduleIndex]
-          : undefined;
-
-        if (result !== undefined) {
-          module.exports = result;
-        } else {
-          if (exportsObject) {
-            module.exports = exportsObject;
-          } else if (moduleObject) {
-            module.exports = moduleObject.exports;
-          }
-        }
-
-        if (
-          module.exports &&
-          typeof module.exports === "object" &&
-          (
-            (result === undefined && (module.exports === exportsObject ||
-              module.exports === moduleObject?.exports)) ||
-            result === exportsObject ||
-            result === moduleObject?.exports
-          )
-        ) {
-          module.exports = this.freezeResolvedExports(module.exports);
-        }
-        module.resolved = true;
-        return module.exports;
-      } finally {
-        this.loading.delete(id);
-      }
-    }
-
-    private generateAnonymousId(): string {
-      return `__anonymous_${Date.now()}_${
-        Math.random().toString(36).substr(2, 9)
-      }`;
-    }
-  }
-
-  const loader = new AMDLoader();
-  return {
-    define: loader.define.bind(loader),
-    require: loader.require.bind(loader),
-  };
-}
