--- packages/js-compiler/typescript/bundler/mod.ts
@@ -1,1 +1,0 @@
-export { bundleAMDOutput, type BundleAMDOutputConfig } from "./bundle.ts";
