+++ packages/runner/deno.json
@@ -24,7 +24,6 @@
     "./storage/cache": "./src/storage/cache.ts",
     "./storage/cache.deno": "./src/storage/cache.deno.ts",
     "./cfc": "./src/cfc/mod.ts",
-    "./cfc/label-view-core": "./src/cfc/label-view-core.ts",
-    "./compilation-cache": "./src/compilation-cache/mod.ts"
+    "./cfc/label-view-core": "./src/cfc/label-view-core.ts"
   }
 }
