+++ packages/runner/src/pattern-manager.ts
@@ -21,7 +21,6 @@ import { createRef } from "./create-ref.ts";
 import type {
   CacheableModule,
   CompiledModuleArtifact,
-  CompileResult,
   EvaluateResult,
   Exports,
 } from "./harness/types.ts";
@@ -548,61 +547,22 @@ export class PatternManager {
       program = input;
     }
 
-    // ESM module-record loader path (experimental, default off).
-    if (
-      this.runtime.experimental.esmModuleLoader === true &&
-      this.runtime.harness.compileToRecordGraph &&
-      this.runtime.harness.evaluateRecordGraph
-    ) {
-      // Use the content-addressed cell cache when we have a target space and
-      // CFC is enforced (the compiled-set integrity label only persists — and
-      // is only trusted on read — under an enforcing mode; see cell-cache).
-      if (cacheCtx && this.runtime.cfcEnforcementMode !== "disabled") {
-        return await this.compileViaCellCache(program, cacheCtx);
-      }
-      const { id, graph, mainSpecifier, entryIdentity } = await this.runtime
-        .harness.compileToRecordGraph(program);
-      cacheCtx?.onEntryIdentity?.(entryIdentity);
-      const result = this.runtime.harness.evaluateRecordGraph(
-        id,
-        graph,
-        mainSpecifier,
-        program.files,
-      );
-      return this.patternFromEvaluation(result, program);
+    // Use the content-addressed cell cache when we have a target space and
+    // CFC is enforced (the compiled-set integrity label only persists — and
+    // is only trusted on read — under an enforcing mode; see cell-cache).
+    if (cacheCtx && this.runtime.cfcEnforcementMode !== "disabled") {
+      return await this.compileViaCellCache(program, cacheCtx);
     }
-
-    const { cachedCompiler } = this.runtime;
-    if (cachedCompiler) {
-      const programHash = createRef(
-        { src: program },
-        "pattern source",
-      ).toString();
-      let compileResult = await cachedCompiler.get(programHash);
-      let loadedFromCache = true;
-      if (!compileResult) {
-        loadedFromCache = false;
-        compileResult = await this.runtime.harness.compile(program);
-        // Fire-and-forget cache write — does not block the load. Tracked in
-        // compileCacheWrites so flushCompileCacheWrites() can await it (graceful
-        // shutdown / deterministic tests), matching the ESM-path write-backs.
-        // Without this, a subsequent load of the same program can read the cache
-        // before this write lands and recompile (the piece.test.ts
-        // "0 in-client compilations" flake).
-        const cacheWrite = cachedCompiler.set(programHash, compileResult).catch(
-          () => {},
-        );
-        this.compileCacheWrites.add(cacheWrite);
-        cacheWrite.finally(() => this.compileCacheWrites.delete(cacheWrite));
-      }
-      return this.evaluateToPattern(compileResult, program, {
-        skipBundleValidation: loadedFromCache,
-      });
-    }
-
-    // No persistent cache — compile and evaluate directly
-    const compileResult = await this.runtime.harness.compile(program);
-    return this.evaluateToPattern(compileResult, program);
+    const { id, graph, mainSpecifier, entryIdentity } = await this.runtime
+      .harness.compileToRecordGraph(program);
+    cacheCtx?.onEntryIdentity?.(entryIdentity);
+    const result = this.runtime.harness.evaluateRecordGraph(
+      id,
+      graph,
+      mainSpecifier,
+      program.files,
+    );
+    return this.patternFromEvaluation(result, program);
   }
 
   /**
@@ -809,8 +769,8 @@ export class PatternManager {
    * recovery, which the caller handles by falling back to the patternId load.
    *
    * Returns the pattern, or `undefined` when the by-identity load is
-   * unavailable (ESM loader off / CFC not enforcing / closure absent or
-   * incomplete / invalid) so the caller can fall back to `loadPattern`.
+   * unavailable (CFC not enforcing / closure absent or incomplete / invalid)
+   * so the caller can fall back to `loadPattern`.
    */
   async loadPatternByIdentity(
     entryIdentity: string,
@@ -818,11 +778,7 @@ export class PatternManager {
     space: MemorySpace,
   ): Promise<Pattern | undefined> {
     const harness = this.runtime.harness;
-    if (
-      this.runtime.experimental.esmModuleLoader !== true ||
-      !harness.evaluateCachedModules ||
-      this.runtime.cfcEnforcementMode === "disabled"
-    ) {
+    if (this.runtime.cfcEnforcementMode === "disabled") {
       return undefined;
     }
     // In-memory fast path (CT-1623): the module may already be live from a
@@ -1115,21 +1071,7 @@ export class PatternManager {
     }
   }
 
-  private async evaluateToPattern(
-    { id, jsScript }: CompileResult,
-    program: RuntimeProgram,
-    options?: { skipBundleValidation?: boolean },
-  ): Promise<Pattern> {
-    const result = await this.runtime.harness.evaluate(
-      id,
-      jsScript,
-      program.files,
-      options,
-    );
-    return this.patternFromEvaluation(result, program);
-  }
-
-  // Resolve a Pattern from an evaluate result (shared by the AMD and ESM paths).
+  // Resolve a Pattern from an evaluate result.
   private patternFromEvaluation(
     result: EvaluateResult,
     program: RuntimeProgram,
