+++ packages/runner/src/runtime.ts
@@ -24,11 +24,6 @@ import {
   resetPersistentSchedulerStateConfig,
   setPersistentSchedulerStateConfig,
 } from "@commonfabric/memory/v2";
-import {
-  getEsmModuleLoaderConfig,
-  resetEsmModuleLoaderConfig,
-  setEsmModuleLoaderConfig,
-} from "./sandbox/esm-loader-config.ts";
 import { PatternEnvironment, setPatternEnvironment } from "./builder/env.ts";
 import { AsyncSemaphoreQueue, type QueueConfig } from "./queue.ts";
 import type {
@@ -77,7 +72,6 @@ import { isDeno } from "@commonfabric/utils/env";
 import { popFrame, pushFrame } from "./builder/pattern.ts";
 import type { Frame } from "./builder/types.ts";
 import type { ConsoleMessage } from "./interface.ts";
-import type { CachedCompiler } from "./compilation-cache/mod.ts";
 import type {
   WriteStackTraceEntry,
   WriteStackTraceMatcher,
@@ -186,13 +180,6 @@ export interface ExperimentalOptions {
   persistentSchedulerState?: boolean | undefined;
   /** Preserve cumulative scheduler write history instead of using current-known writes. */
   schedulerHistoricalMightWrite?: boolean | undefined;
-  /**
-   * Load compiled patterns as a graph of per-module records through the SES
-   * Compartment module system (synchronous `importNow`) instead of evaluating
-   * one flattened AMD bundle. Default off; the AMD path remains the default.
-   * See docs/specs/module-loading.md (Phases 2–5).
-   */
-  esmModuleLoader?: boolean | undefined;
 }
 
 export interface RuntimeOptions {
@@ -211,9 +198,6 @@ export interface RuntimeOptions {
   cfcEnforcementMode?: CfcEnforcementMode;
   /** Deterministic provider for the trust snapshot attached to each new tx. */
   trustSnapshotProvider?: () => TrustSnapshot | undefined;
-  /** Optional compilation cache for persistent caching of compiled JS.
-   *  If absent, no persistent caching is performed (same as before). */
-  cachedCompiler?: CachedCompiler;
   /** Replace runner-owned frames with `<CF_INTERNAL>` in surfaced stacks. */
   hideInternalStackFrames?: boolean;
 }
@@ -322,7 +306,6 @@ export class Runtime {
   readonly storageManager: IStorageManager;
   readonly trustSnapshotProvider: () => TrustSnapshot | undefined;
   readonly telemetry: RuntimeTelemetry;
-  readonly cachedCompiler?: CachedCompiler;
   /** Resolved experimental flags (all properties present, defaulting to `false`). */
   readonly experimental: ExperimentalOptions;
   readonly apiUrl: URL;
@@ -339,7 +322,6 @@ export class Runtime {
       modernCellRep: undefined,
       persistentSchedulerState: undefined,
       schedulerHistoricalMightWrite: undefined,
-      esmModuleLoader: undefined,
       ...options.experimental,
     };
 
@@ -365,9 +347,6 @@ export class Runtime {
     );
     this.experimental.persistentSchedulerState =
       getPersistentSchedulerStateConfig();
-    // Env-seeded default (CF_ESM_MODULE_LOADER) unless an explicit option is set.
-    setEsmModuleLoaderConfig(this.experimental.esmModuleLoader);
-    this.experimental.esmModuleLoader = getEsmModuleLoaderConfig();
 
     this.id = options.storageManager.id;
     this.apiUrl = new URL(options.apiUrl);
@@ -376,7 +355,6 @@ export class Runtime {
       : new StaticCacheHTTP(new URL("/static", this.apiUrl));
 
     this.telemetry = options.telemetry ?? new RuntimeTelemetry();
-    this.cachedCompiler = options.cachedCompiler;
 
     // Create harness first (no dependencies on other services)
     this.harness = new Engine(this, {
@@ -424,13 +402,6 @@ export class Runtime {
     // fallback in builder/env.ts. This is still a singleton. TODO(seefeld).
     if (options.patternEnvironment) {
       setPatternEnvironment(options.patternEnvironment);
-    }
-
-    // Fire-and-forget startup eviction for compilation cache
-    if (this.cachedCompiler) {
-      this.cachedCompiler.evictStale().catch((err) => {
-        console.warn("Compilation cache eviction failed:", err);
-      });
     }
 
     if (options.debug) {
@@ -551,7 +522,6 @@ export class Runtime {
     // Reset experimental config to defaults.
     resetModernCellRepConfig();
     resetPersistentSchedulerStateConfig();
-    resetEsmModuleLoaderConfig();
 
     // Clear the current runtime reference
     // Removed setCurrentRuntime call - no longer using singleton pattern
