--- packages/runner/test/git-fingerprint.test.ts
@@ -1,55 +1,0 @@
-import { describe, it } from "@std/testing/bdd";
-import { expect } from "@std/expect";
-import { computeGitFingerprint } from "../src/compilation-cache/git-fingerprint.ts";
-
-describe("computeGitFingerprint", () => {
-  it("returns a string in a git repository", async () => {
-    const fingerprint = await computeGitFingerprint();
-
-    // We're running inside a git repo, so this should succeed
-    expect(fingerprint).toBeDefined();
-    expect(typeof fingerprint).toBe("string");
-    // Clean tree → raw HEAD SHA (40 hex chars), dirty → sha256 (64 hex chars)
-    expect([40, 64]).toContain(fingerprint!.length);
-  });
-
-  it("returns the same value on consecutive calls (clean tree)", async () => {
-    const first = await computeGitFingerprint();
-    const second = await computeGitFingerprint();
-
-    expect(first).toBe(second);
-  });
-
-  it("explicit SHA is returned as-is", async () => {
-    const sha = "abc123def456";
-    const fingerprint = await computeGitFingerprint(sha);
-
-    expect(fingerprint).toBe(sha);
-  });
-
-  it("explicit SHA takes priority over git", async () => {
-    const fromGit = await computeGitFingerprint();
-    const fromExplicit = await computeGitFingerprint("abc123");
-
-    // We're in a git repo, but explicit SHA should still win
-    expect(fromExplicit).toBe("abc123");
-    expect(fromExplicit).not.toBe(fromGit);
-  });
-
-  it("different explicit SHAs produce different fingerprints", async () => {
-    const a = await computeGitFingerprint("sha-aaa");
-    const b = await computeGitFingerprint("sha-bbb");
-
-    expect(a).not.toBe(b);
-  });
-
-  it("returns undefined with no git and no explicit SHA", async () => {
-    // Can't easily remove git from the test environment, but we can
-    // verify the contract: no explicitSha + git failure → undefined.
-    // The "returns a string in a git repository" test proves the git
-    // path works; the "explicit SHA is returned as-is" test proves
-    // the explicit path works. The undefined path is the catch block
-    // when both are absent — exercised in Docker deploys without
-    // .git or TOOLSHED_GIT_SHA.
-  });
-});
