+++ packages/runtime-client/runtime-client.ts
@@ -95,7 +95,6 @@ export class RuntimeClient extends EventEmitter<RuntimeClientEvents> {
       experimental: options.experimental,
       cfcEnforcementMode: options.cfcEnforcementMode,
       trustSnapshot: options.trustSnapshot,
-      buildHash: options.buildHash,
     });
     return new RuntimeClient(initialized, options);
   }
