+++ packages/shell/felt.config.ts
@@ -47,14 +47,6 @@ const config: Config = {
       "$EXPERIMENTAL_PERSISTENT_SCHEDULER_STATE": Deno.env.get(
         "EXPERIMENTAL_PERSISTENT_SCHEDULER_STATE",
       ),
-      // The ESM module-record loader shares one env var with the server/runner
-      // side (`CF_ESM_MODULE_LOADER`), unlike the `EXPERIMENTAL_*`-named flags
-      // above, so the same var toggles it for both the Deno runtime and the
-      // browser worker.
-      "$EXPERIMENTAL_ESM_MODULE_LOADER": Deno.env.get("CF_ESM_MODULE_LOADER"),
-      "$COMPILATION_CACHE_CLIENT": Deno.env.get(
-        "COMPILATION_CACHE_CLIENT",
-      ) ?? "true",
     },
     supported: {
       // Provide polyfills for `using` resource management
